Modify output of button event handler
The logic of how the click event is handled is updated to ensure the
user input an integer between 1 and 100.

Once the user enters a valid input, a for loop is used to select all the
square div elements and delete them one by one.

The createGrid function is called after resetting the container.