Move the event handler for changing color
The event handler that handles the mouseover event to change color is
moved inside the createGrid function. This is because the createGrid
function is called after DOM content is loaded.

Hence the square elements could not be referred to before the grid is
created. Moving the event handler inside the createGrid function allows
all the squares to be selected and manipulated after the grid is created.